+++ src/laybasic/laybasic/layLineStyles.cc
@@ -520,13 +520,13 @@ LineStyles::merge (const LineStyles &other, std::map<unsigned int, unsigned int>
 {
   //  insert the standard pattern into the map (for completeness)
   for (iterator c = begin (); c != begin_custom (); ++c) {
-    index_map.insert (std::make_pair (std::distance (begin (), c), std::distance (begin (), c)));
+    index_map.insert (std::make_pair ((unsigned int) std::distance (begin (), c), (unsigned int) std::distance (begin (), c)));
   }
 
   //  build an index of present pattern
   std::map <LineStyleInfo, unsigned int, style_less_f> styles;
   for (iterator c = begin_custom (); c != end (); ++c) {
-    styles.insert (std::make_pair (*c, std::distance (begin (), c)));
+    styles.insert (std::make_pair (*c, (unsigned int) std::distance (begin (), c)));
   }
 
   //  map the pattern of other into *this, possibly creating new ones
@@ -539,7 +539,7 @@ LineStyles::merge (const LineStyles &other, std::map<unsigned int, unsigned int>
     } else {
       new_index = p->second;
     }
-    index_map.insert (std::make_pair (std::distance (other.begin (), c), new_index));
+    index_map.insert (std::make_pair ((unsigned int) std::distance (other.begin (), c), new_index));
   }
 }
